`@remotion/google-fonts`: New `loadFontFromInfo()` API
