+++ packages/docs/docs/google-fonts/index.mdx
@@ -7,7 +7,7 @@ title: '@remotion/google-fonts'
 
 import Tabs from '@theme/Tabs';
 import TabItem from '@theme/TabItem';
-import {TableOfContents} from '../../components/TableOfContents/google-fonts';
+import {TableOfContents} from './TableOfContents';
 
 The `@remotion/google-fonts` package provides APIs for easily integrating [Google Fonts](https://fonts.google.com/) into Remotion.
