+++ packages/example/src/LoadFontBasedOnInfo.ts
@@ -1,7 +1,8 @@
 import {getInfo} from '@remotion/google-fonts/InterTight';
-import {loadFontFromInfo} from '@remotion/google-fonts/load-font-from-info';
+import {loadFontFromInfo} from '@remotion/google-fonts/from-info';
 
 const info = getInfo();
 const {fontFamily, waitUntilDone} = loadFontFromInfo(info, 'italic');
 
+console.log(fontFamily);
 waitUntilDone();
